+++ lib/utils/get_random_string.dart
@@ -1,0 +1,9 @@
+import 'dart:math';
+
+String getRandomString(int length) {
+  const characters =
+      '+-*=?AaBbCcDdEeFfGgHhIiJjKkLlMmNnOoPpQqRrSsTtUuVvWwXxYyZz';
+  Random random = Random();
+  return String.fromCharCodes(Iterable.generate(
+      length, (_) => characters.codeUnitAt(random.nextInt(characters.length))));
+}